ref(explore): Share group by validation helpers
Centralize invalid group by cleanup, visibility, and validated tag merging across spans and metrics while preserving their mode-specific behavior.

Refs EXP-1024
