+++ static/app/views/explore/metrics/metricToolbar/groupBySelector.tsx
@@ -1,6 +1,5 @@
-import {useCallback, useLayoutEffect, useMemo, useRef} from 'react';
+import {useCallback, useMemo} from 'react';
 import {useQuery} from '@tanstack/react-query';
-import isEqual from 'lodash/isEqual';
 
 import type {SelectOption} from '@sentry/scraps/compactSelect';
 import {CompactSelect} from '@sentry/scraps/compactSelect';
@@ -8,12 +7,10 @@ import {OverlayTrigger} from '@sentry/scraps/overlayTrigger';
 
 import {usePageFilters} from 'sentry/components/pageFilters/usePageFilters';
 import {t} from 'sentry/locale';
-import type {TagCollection} from 'sentry/types/group';
-import {FieldKind} from 'sentry/utils/fields';
 import {useOrganization} from 'sentry/utils/useOrganization';
-import {prettifyAttributeName} from 'sentry/views/explore/components/traceItemAttributes/utils';
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
 import {useGroupByFields} from 'sentry/views/explore/hooks/useGroupByFields';
+import {useValidatedGroupBys} from 'sentry/views/explore/hooks/useValidatedGroupBys';
 import {HiddenTraceMetricGroupByFields} from 'sentry/views/explore/metrics/constants';
 import {useValidateMetricsTab} from 'sentry/views/explore/metrics/hooks/useValidateMetricsTab';
 import type {TraceMetric} from 'sentry/views/explore/metrics/metricQuery';
@@ -23,12 +20,12 @@ import {
   useSetQueryParamsGroupBys,
 } from 'sentry/views/explore/queryParams/context';
 import {TraceItemDataset} from 'sentry/views/explore/types';
+import {mergeValidatedGroupByTags} from 'sentry/views/explore/utils/groupByValidation';
 import {sortSearchedAttributes} from 'sentry/views/explore/utils/sortSearchedAttributes';
 import {
   selectTraceItemTagCollection,
   traceItemAttributeKeysOptions,
 } from 'sentry/views/explore/utils/traceItemAttributeKeysOptions';
-import type {EventValidationData} from 'sentry/views/explore/utils/validateEventParamsOptions';
 
 interface GroupBySelectorProps {
   /**
@@ -69,78 +66,14 @@ export function GroupBySelector({
     isLoading: validationLoading,
     isPlaceholderData: validationIsPlaceholderData,
   } = useValidateMetricsTab();
-  const pendingValidatedGroupBys = useRef<{
-    from: readonly string[];
-    to: readonly string[];
-  } | null>(null);
-  const validationGroupBys = useRef<{
-    data: EventValidationData;
-    groupBys: readonly string[];
-  } | null>(null);
   const validationIsPending =
     validationFetching || validationLoading || validationIsPlaceholderData;
-
-  const validatedGroupBys = useMemo(
-    () => filterInvalidGroupBys(groupBys, validatedSearchQueryData?.field),
-    [groupBys, validatedSearchQueryData?.field]
-  );
-  const visibleGroupBys = useMemo(
-    () =>
-      filterVisibleGroupBys(
-        groupBys,
-        validatedSearchQueryData?.field,
-        validationIsPending
-      ),
-    [groupBys, validatedSearchQueryData?.field, validationIsPending]
-  );
-
-  useLayoutEffect(() => {
-    if (pendingValidatedGroupBys.current) {
-      if (isEqual(groupBys, pendingValidatedGroupBys.current.to)) {
-        pendingValidatedGroupBys.current = null;
-      } else if (
-        isEqual(groupBys, pendingValidatedGroupBys.current.from) &&
-        isEqual(validatedGroupBys, pendingValidatedGroupBys.current.to)
-      ) {
-        return;
-      }
-    }
-
-    if (validationIsPending || !validatedSearchQueryData) {
-      return;
-    }
-
-    let validationGroupBySnapshot = validationGroupBys.current;
-    if (
-      !validationGroupBySnapshot?.data ||
-      validationGroupBySnapshot.data !== validatedSearchQueryData
-    ) {
-      validationGroupBySnapshot = {
-        data: validatedSearchQueryData,
-        groupBys,
-      };
-      validationGroupBys.current = validationGroupBySnapshot;
-    }
-
-    if (
-      !isEqual(groupBys, validationGroupBySnapshot.groupBys) ||
-      isEqual(groupBys, validatedGroupBys)
-    ) {
-      return;
-    }
-
-    pendingValidatedGroupBys.current = {
-      from: groupBys,
-      to: validatedGroupBys,
-    };
-    setGroupBys(validatedGroupBys);
-  }, [
+  const {visibleGroupBys} = useValidatedGroupBys({
     groupBys,
-    setGroupBys,
-    validatedGroupBys,
-    validatedSearchQueryData,
+    validationData: validatedSearchQueryData,
     validationIsPending,
-  ]);
+    onGroupBysCleanup: setGroupBys,
+  });
 
   const traceMetricFilter = createTraceMetricFilter(traceMetric);
 
@@ -172,7 +105,7 @@ export function GroupBySelector({
       )
     );
 
-    return mergeValidatedTags({
+    return mergeValidatedGroupByTags({
       booleanTags: visibleBooleanTags,
       numberTags: visibleNumberTags,
       stringTags: visibleStringTags,
@@ -247,67 +180,4 @@ export function GroupBySelector({
       style={{width: '100%'}}
     />
   );
-}
-
-function filterInvalidGroupBys(
-  groupBys: readonly string[],
-  fields: EventValidationData['field'] | undefined
-): string[] {
-  const invalidFields = new Set(
-    fields?.filter(field => !field.valid).map(field => field.name)
-  );
-
-  if (invalidFields.size === 0) {
-    return [...groupBys];
-  }
-
-  return groupBys.filter(groupBy => groupBy === '' || !invalidFields.has(groupBy));
-}
-
-function filterVisibleGroupBys(
-  groupBys: readonly string[],
-  fields: EventValidationData['field'] | undefined,
-  validationIsPending: boolean
-): string[] {
-  return groupBys.filter(groupBy => {
-    if (groupBy === '') {
-      return true;
-    }
-
-    const field = fields?.find(({name}) => name === groupBy);
-    return field?.valid || (!validationIsPending && field?.valid !== false);
-  });
-}
-
-function mergeValidatedTags({
-  booleanTags,
-  numberTags,
-  stringTags,
-  validatedFields = [],
-}: {
-  booleanTags: TagCollection;
-  numberTags: TagCollection;
-  stringTags: TagCollection;
-  validatedFields?: EventValidationData['field'];
-}) {
-  const validatedBooleanTags = {...booleanTags};
-  const validatedNumberTags = {...numberTags};
-  const validatedStringTags = {...stringTags};
-
-  for (const validatedField of validatedFields) {
-    const tag = {
-      key: validatedField.name,
-      name: prettifyAttributeName(validatedField.name),
-    };
-
-    if (validatedField.attrType === 'boolean') {
-      validatedBooleanTags[validatedField.name] = {...tag, kind: FieldKind.BOOLEAN};
-    } else if (validatedField.attrType === 'number') {
-      validatedNumberTags[validatedField.name] = {...tag, kind: FieldKind.MEASUREMENT};
-    } else if (validatedField.attrType === 'string') {
-      validatedStringTags[validatedField.name] = {...tag, kind: FieldKind.TAG};
-    }
-  }
-
-  return {validatedBooleanTags, validatedNumberTags, validatedStringTags};
 }
