+++ static/app/views/explore/toolbar/toolbarGroupBy.tsx
@@ -1,8 +1,5 @@
-import {useCallback, useLayoutEffect, useMemo, useRef, useState} from 'react';
-import isEqual from 'lodash/isEqual';
+import {useCallback, useMemo, useState} from 'react';
 
-import type {TagCollection} from 'sentry/types/group';
-import {FieldKind} from 'sentry/utils/fields';
 import {useDebouncedValue} from 'sentry/utils/useDebouncedValue';
 import {
   ToolbarFooter,
@@ -13,14 +10,18 @@ import {
   ToolbarGroupByDropdown,
   ToolbarGroupByHeader,
 } from 'sentry/views/explore/components/toolbar/toolbarGroupBy';
-import {prettifyAttributeName} from 'sentry/views/explore/components/traceItemAttributes/utils';
 import {DragNDropContext} from 'sentry/views/explore/contexts/dragNDropContext';
 import {Mode} from 'sentry/views/explore/contexts/pageParamsContext/mode';
 import type {Column} from 'sentry/views/explore/hooks/useDragNDropColumns';
 import {useGroupByFields} from 'sentry/views/explore/hooks/useGroupByFields';
 import {useSpanItemAttributes} from 'sentry/views/explore/hooks/useTraceItemAttributes';
+import {useValidatedGroupBys} from 'sentry/views/explore/hooks/useValidatedGroupBys';
 import {useValidateSpansTab} from 'sentry/views/explore/spans/hooks/useValidateSpansTab';
 import {TraceItemDataset} from 'sentry/views/explore/types';
+import {
+  mergeValidatedGroupByTags,
+  shouldHideGroupByForValidation,
+} from 'sentry/views/explore/utils/groupByValidation';
 import type {EventValidationData} from 'sentry/views/explore/utils/validateEventParamsOptions';
 
 interface ToolbarGroupByProps {
@@ -35,83 +36,25 @@ export function ToolbarGroupBy({groupBys, setGroupBys}: ToolbarGroupByProps) {
     isLoading: validationLoading,
     isPlaceholderData: validationIsPlaceholderData,
   } = useValidateSpansTab();
-  const pendingValidatedGroupBys = useRef<{
-    from: readonly string[];
-    to: readonly string[];
-  } | null>(null);
-  const validationGroupBys = useRef<{
-    data: EventValidationData;
-    groupBys: readonly string[];
-  } | null>(null);
   const validationIsPending =
     validationFetching || validationLoading || validationIsPlaceholderData;
 
-  const validatedGroupBys = useMemo(
-    () => filterInvalidGroupBys(groupBys, validatedSearchQueryData?.field),
-    [groupBys, validatedSearchQueryData?.field]
-  );
-  const visibleGroupBys = useMemo(
-    () =>
-      filterVisibleGroupBys(
-        groupBys,
-        validatedSearchQueryData?.field,
-        validationIsPending
-      ),
-    [groupBys, validatedSearchQueryData?.field, validationIsPending]
-  );
-
-  useLayoutEffect(() => {
-    if (pendingValidatedGroupBys.current) {
-      if (isEqual(groupBys, pendingValidatedGroupBys.current.to)) {
-        pendingValidatedGroupBys.current = null;
-      } else if (
-        isEqual(groupBys, pendingValidatedGroupBys.current.from) &&
-        isEqual(validatedGroupBys, pendingValidatedGroupBys.current.to)
-      ) {
-        return;
+  const cleanupInvalidGroupBys = useCallback(
+    (validatedGroupBys: string[]) => {
+      if (validatedGroupBys.some(Boolean)) {
+        setGroupBys(validatedGroupBys);
+      } else {
+        setGroupBys(validatedGroupBys, Mode.SAMPLES);
       }
-    }
-
-    if (validationIsPending || !validatedSearchQueryData) {
-      return;
-    }
-
-    let validationGroupBySnapshot = validationGroupBys.current;
-    if (
-      !validationGroupBySnapshot?.data ||
-      validationGroupBySnapshot.data !== validatedSearchQueryData
-    ) {
-      validationGroupBySnapshot = {
-        data: validatedSearchQueryData,
-        groupBys,
-      };
-      validationGroupBys.current = validationGroupBySnapshot;
-    }
-
-    if (
-      !isEqual(groupBys, validationGroupBySnapshot.groupBys) ||
-      isEqual(groupBys, validatedGroupBys)
-    ) {
-      return;
-    }
-
-    pendingValidatedGroupBys.current = {
-      from: groupBys,
-      to: validatedGroupBys,
-    };
-
-    if (validatedGroupBys.some(Boolean)) {
-      setGroupBys(validatedGroupBys);
-    } else {
-      setGroupBys(validatedGroupBys, Mode.SAMPLES);
-    }
-  }, [
+    },
+    [setGroupBys]
+  );
+  const {visibleGroupBys} = useValidatedGroupBys({
     groupBys,
-    setGroupBys,
-    validatedGroupBys,
-    validatedSearchQueryData,
+    validationData: validatedSearchQueryData,
     validationIsPending,
-  ]);
+    onGroupBysCleanup: cleanupInvalidGroupBys,
+  });
 
   const setGroupBysWithOp = useCallback(
     (columns: string[], op: 'insert' | 'update' | 'delete' | 'reorder') => {
@@ -207,7 +150,12 @@ function ToolbarGroupByItem({
       };
     }
 
-    return mergeValidatedTags({booleanTags, numberTags, stringTags, validatedField});
+    return mergeValidatedGroupByTags({
+      booleanTags,
+      numberTags,
+      stringTags,
+      validatedFields: [validatedField],
+    });
   }, [booleanTags, column, numberTags, stringTags, validatedSearchQueryData?.field]);
 
   const options = useGroupByFields({
@@ -241,116 +189,4 @@ function ToolbarGroupByItem({
       onColumnDelete={onColumnDelete}
     />
   );
-}
-
-function filterInvalidGroupBys(
-  groupBys: readonly string[],
-  fields: EventValidationData['field'] | undefined
-): string[] {
-  const invalidFields = new Set(
-    fields?.filter(field => !field.valid).map(field => field.name)
-  );
-
-  if (invalidFields.size === 0) {
-    return [...groupBys];
-  }
-
-  return groupBys.filter(groupBy => groupBy === '' || !invalidFields.has(groupBy));
-}
-
-function filterVisibleGroupBys(
-  groupBys: readonly string[],
-  fields: EventValidationData['field'] | undefined,
-  validationIsPending: boolean
-): string[] {
-  return groupBys.filter(
-    groupBy => !shouldHideGroupByForValidation(groupBy, fields, validationIsPending)
-  );
-}
-
-function shouldHideGroupByForValidation(
-  groupBy: string,
-  fields: EventValidationData['field'] | undefined,
-  validationIsPending: boolean
-): boolean {
-  if (groupBy === '') {
-    return false;
-  }
-
-  const field = fields?.find(({name}) => name === groupBy);
-
-  if (field?.valid) {
-    return false;
-  }
-
-  return validationIsPending || field?.valid === false;
-}
-
-function mergeValidatedTags({
-  booleanTags,
-  numberTags,
-  stringTags,
-  validatedField,
-}: {
-  booleanTags: TagCollection;
-  numberTags: TagCollection;
-  stringTags: TagCollection;
-  validatedField: EventValidationData['field'][number];
-}) {
-  switch (validatedField.attrType) {
-    case 'boolean': {
-      const validatedBooleanTags = {
-        ...booleanTags,
-        [validatedField.name]: {
-          key: validatedField.name,
-          name: prettifyAttributeName(validatedField.name),
-          kind: FieldKind.BOOLEAN,
-        },
-      };
-
-      return {
-        validatedBooleanTags,
-        validatedNumberTags: numberTags,
-        validatedStringTags: stringTags,
-      };
-    }
-    case 'number': {
-      const validatedNumberTags = {
-        ...numberTags,
-        [validatedField.name]: {
-          key: validatedField.name,
-          name: prettifyAttributeName(validatedField.name),
-          kind: FieldKind.MEASUREMENT,
-        },
-      };
-
-      return {
-        validatedBooleanTags: booleanTags,
-        validatedNumberTags,
-        validatedStringTags: stringTags,
-      };
-    }
-    case 'string': {
-      const validatedStringTags = {
-        ...stringTags,
-        [validatedField.name]: {
-          key: validatedField.name,
-          name: prettifyAttributeName(validatedField.name),
-          kind: FieldKind.TAG,
-        },
-      };
-
-      return {
-        validatedBooleanTags: booleanTags,
-        validatedNumberTags: numberTags,
-        validatedStringTags,
-      };
-    }
-    default:
-      return {
-        validatedBooleanTags: booleanTags,
-        validatedNumberTags: numberTags,
-        validatedStringTags: stringTags,
-      };
-  }
 }
